# backend/openapi/openapi-1.18.0-beta/openapi-1.18.0-beta.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cyber Media Center API (赛博影视)",
    "description": "## 专业发烧级影视中心后端接口文档\n\n此文档定义了 Cyber Media 的核心 API 交互规范。旨在取代 Plex/Emby，提供高性能、低延迟的媒体管理与流媒体服务。\n\n### 🔗 API 生命周期说明 (x-lifecycle-status)\n我们在每个接口中定义了开发状态，请前端同学注意：\n\n1.  🔴 **pending** (未开发): 接口仅有定义，后端尚未实现，调用会返回 501。\n2.  🟠 **implemented_unverified** (已开发未验证): 代码已合并，但未经完整测试，可能存在 Bug。\n3.  ❌ **failed_verification** (验证失败): 测试发现严重问题，暂时禁止使用。\n4.  ⚠️ **passed_imperfect** (验证通过但格式不规范): 功能可用，但返回字段可能不全或命名不符合规范，后续会变动。\n5.  ✅ **stable** (验证通过): 接口稳定，可放心集成。\n\n### 🛡️ 鉴权说明\n所有非公开接口需在 Header 中携带 `Authorization: Bearer <token>`。",
    "version": "1.18.0-beta",
    "contact": {
      "name": "Cyber Media Core Team",
      "email": "dev@cybermedia.local"
    }
  },
  "servers": [
    {
      "url": "https://pw.pioneer.fan:84",
      "description": "生产环境 (Production HTTPS)"
    },
    {
      "url": "http://pioneer.fan:84",
      "description": "生产环境 (Production HTTP IPv4)"
    },
    {
      "url": "http://localhost:5004",
      "description": "本地开发环境"
    }
  ],
  "paths": {
...
    "/api/v1/user/history": {
      "get": {
        "summary": "获取观看历史",
        "description": "分页获取观看历史记录，用于历史页面和续播入口。历史接口保留，但返回的播放上下文不再包含 is_played 标记。",
        "operationId": "getHistory",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "历史记录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "上报播放进度",
        "description": "上报播放进度心跳。使用独立音频实时转码时，该接口也是后端 watchdog 的保活信号；超过音频转码响应头 X-Cyber-History-Timeout 指定时间未提交对应 resource_id 的进度时，后端会停止转码。",
        "operationId": "reportProgress",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "resource_id",
                  "position_sec",
                  "total_duration"
                ],
                "properties": {
                  "resource_id": {
                    "type": "string"
                  },
                  "position_sec": {
                    "type": "integer",
                    "description": "当前播放秒数"
                  },
                  "total_duration": {
                    "type": "integer",
                    "description": "视频总秒数"
                  },
                  "device_id": {
                    "type": "string",
                    "description": "设备唯一标识，可选"
                  },
                  "device_name": {
                    "type": "string",
                    "description": "设备名称，可选"
                  },
                  "session_id": {
                    "type": "string",
                    "description": "播放会话 ID，可选。与 audio-transcode 的 session_id 一致时可精确刷新对应转码流；不传时会刷新该资源下所有活跃转码流。",
                    "pattern": "^[A-Za-z0-9_.:-]{1,80}$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上报成功"
          }
        }
      },
      "delete": {
        "summary": "清空观看历史",
        "description": "清空所有观看历史记录。",
        "operationId": "clearAllHistory",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "清空成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/history/{resource_id}": {
      "delete": {
        "summary": "删除单条历史记录",
        "description": "删除指定资源的观看历史记录。",
        "operationId": "deleteHistoryItem",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "description": "要删除的资源 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "记录不存在"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "msg": {
            "type": "string",
            "example": "success"
          },
          "trace_id": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
    }
  }
}